+++ src/aui-modal/js/aui-modal.js
@@ -75,6 +75,7 @@ A.Modal = A.Base.create(MODAL, A.Widget, [
 
         var eventHandles = [
             A.after(instance._afterFillHeight, instance, FILL_HEIGHT),
+            A.before(instance._beforeFillHeight, instance, FILL_HEIGHT),
             instance.after('resize:end', A.bind(instance._syncResizeDimensions, instance)),
             instance.after(DRAGGABLE_CHANGE, instance._afterDraggableChange),
             instance.after(RESIZABLE_CHANGE, instance._afterResizableChange),
@@ -201,6 +202,18 @@ A.Modal = A.Base.create(MODAL, A.Widget, [
             instance._userInteractionHandle = instance.once(
                 [CLICK, MOUSEMOVE], instance._onUserInitInteraction, instance);
         }
+    },
+
+    /**
+     * Fire before modal body height is set.
+     *
+     * @method _beforeFillHeight
+     * @protected
+     */
+    _beforeFillHeight: function() {
+        var instance = this;
+
+        instance._fillMaxHeight('none');
     },
 
     /**