+++ src/aui-modal/HISTORY.md
@@ -1,2 +1,6 @@
-aui-modal
+# AUI Modal
 ========
+
+## @VERSION@
+
+* [AUI-1909](https://issues.liferay.com/browse/AUI-1909)  Modal max-height prevents modal body offsetHeight from being set correctly